Added partial fragments for graphQL; claims journey update
