+++ Projects/Claims/GraphQL/Octopus/ClaimsFlowAudioRecording.graphql
@@ -3,26 +3,8 @@ mutation ClaimsFlowAudioRecording($input: FlowClaimAudioRecordingInput!, $contex
     context
     currentStep {
       id
-      ... on FlowClaimAudioRecordingStep {
-        id
-        questions
-      }
-      ... on FlowClaimDateOfOccurrencePlusLocationStep {
-        dateOfOccurrenceStep {
-          dateOfOccurrence
-          id
-          maxDate
-        }
-        id
-        locationStep {
-          id
-          location
-          options {
-            displayName
-            value
-          }
-        }
-      }
+      ...FlowClaimAudioRecordingStepFragment
+      ...FlowClaimDateOfOccurrencePlusLocationStepFragment
       ... on FlowClaimDateOfOccurrenceStep {
         dateOfOccurrence
         id
