+++ Projects/Claims/GraphQL/Octopus/ClaimsFlowDateOfOccurrencePlusLocationNext.graphql
@@ -4,26 +4,8 @@ mutation ClaimsFlowDateOfOccurrencePlusLocation($input: FlowClaimDateOfOccurrenc
     id
     currentStep {
       id
-      ... on FlowClaimAudioRecordingStep {
-        id
-        questions
-      }
-      ... on FlowClaimDateOfOccurrencePlusLocationStep {
-        dateOfOccurrenceStep {
-          dateOfOccurrence
-          id
-          maxDate
-        }
-        id
-        locationStep {
-          id
-          location
-          options {
-            displayName
-            value
-          }
-        }
-      }
+      ...FlowClaimAudioRecordingStepFragment
+      ...FlowClaimDateOfOccurrencePlusLocationStepFragment
       ... on FlowClaimDateOfOccurrenceStep {
         dateOfOccurrence
         id
