+++ Projects/Claims/GraphQL/Octopus/ClaimsFlowSingleItem.graphql
@@ -3,26 +3,8 @@ mutation ClaimsFlowSingleItem($context: FlowContext!, $input: FlowClaimSingleIte
     context
     currentStep {
       id
-      ... on FlowClaimAudioRecordingStep {
-        id
-        questions
-      }
-      ... on FlowClaimDateOfOccurrencePlusLocationStep {
-        dateOfOccurrenceStep {
-          dateOfOccurrence
-          maxDate
-          id
-        }
-        id
-        locationStep {
-          id
-          location
-          options {
-            displayName
-            value
-          }
-        }
-      }
+      ...FlowClaimAudioRecordingStepFragment
+      ...FlowClaimDateOfOccurrencePlusLocationStepFragment
       ... on FlowClaimDateOfOccurrenceStep {
         dateOfOccurrence
         id
